tests: a nested lambda's blueprint, asked whether it carries the release tables
Both backends that build a blueprint at a `MakeClosure` are asked for the two
tables an error exit walks, and for the region tables beside them. Each is read
off the same `LirFunction`, so an omission in one is as silent as in the other:
the closure runs, and only an abandoned frame loses anything.

The JIT's blueprint is reached through `translate_function`. `compile` rejects a
function holding a `MakeClosure` before the translator sees it, and the
blueprint is built all the same.

The JIT compiler's tests are split at their four subjects to stay inside the
reading budget: what the solo gate accepts, what a batch accepts, what the
rendered Cranelift IR shows, and what an activation reclaims.
